Basic Model
I added basically everything to run the model. Although I didn't test any of it on the data yet so It will change a lot once we have the datasets inside and cleaned and everything.

But it should in the most basic sense do everything we stated in our plan